requirements.txt
google-genai
opencv-python
pillow
moviepy
requests
schedule
python-dotenv